src/requirements.txt
pandas==2.2.3
psycopg2-binary==2.9.9
redis==5.1.0
websockets==13.1